fix [api]: Implement kv/dump endpoint, use shared kv code
This endpoint was not implemented, or rather, it always returned an
empty object.

This commit also moves the kv code to the shared folder (it used to be
in the api folder) as we will start using some kv features in future
core updates.
