+++ shared/package.json
@@ -29,6 +29,7 @@
     "./errors": "./src/errors.mjs",
     "./fs": "./src/fs.mjs",
     "./logger": "./src/logger.mjs",
+    "./kv": "./src/kv.mjs",
     "./network": "./src/network.mjs",
     "./schema": "./src/schema.mjs",
     "./store": "./src/store.mjs",
@@ -40,6 +41,7 @@
     "glob": "^11.0.0",
     "jsonwebtoken": "^9.0.2",
     "lodash": "^4.17.21",
+    "minimatch": "^10.0.1",
     "node-forge": "^1.3.1",
     "node-jose": "^2.2.0",
     "openpgp": "^5.11.2"
